+++ Classes/Controller/LoginController.php
@@ -23,7 +23,6 @@ use Psr\Http\Message\ServerRequestInterface;
 use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Http\PropagateResponseException;
 use TYPO3\CMS\Core\Http\RedirectResponse;
-use TYPO3\CMS\Core\Http\ServerRequest;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
 
@@ -39,12 +38,7 @@ class LoginController
      */
     public ?ContentObjectRenderer $cObj = null;
 
-    protected ServerRequest $request;
-
-    public function __construct()
-    {
-        $this->request = $GLOBALS['TYPO3_REQUEST'];
-    }
+    protected ServerRequestInterface $request;
 
     public function setContentObjectRenderer(ContentObjectRenderer $cObj): void
     {
@@ -59,22 +53,20 @@ class LoginController
      * If the user has just been logged in and just came back from the authorization server, redirect the user to the
      * final redirect URL.
      *
-     * @param string $_ ignored
-     * @param array|null $pluginConfiguration
      * @throws PropagateResponseException
      */
-    public function login(string $_, ?array $pluginConfiguration): void
+    public function login(string $_, ?array $pluginConfiguration, ServerRequestInterface $request): void
     {
+        $this->request = $request;
         if (is_array($pluginConfiguration)) {
             $this->pluginConfiguration = $pluginConfiguration;
         }
 
-        /** @var Context $context */
         $context = GeneralUtility::makeInstance(Context::class);
         $loginType = $this->request->getParsedBody()['logintype'] ?? $this->request->getQueryParams()['logintype'] ?? '';
         if ($loginType === 'login' || $context->getAspect('frontend.user')->isLoggedIn()) {
             $redirectUrl = $this->determineRedirectUrl();
-            $this->redirect($redirectUrl);
+            throw new PropagateResponseException(new RedirectResponse($redirectUrl));
         }
 
         $authorizationRedirect = $this->getAuthorizationRedirect($this->request, $pluginConfiguration['authorizationUrlOptions.'] ?? []);
@@ -107,13 +99,5 @@ class LoginController
         }
 
         return '/';
-    }
-
-    /**
-     * @throws PropagateResponseException
-     */
-    protected function redirect(string $redirectUrl): void
-    {
-        throw new PropagateResponseException(new RedirectResponse($redirectUrl));
     }
 }